refactor: split pdf reader
